requirements.txt
Flask==3.0.1
flask_wtf==1.2.1
fonttools==4.47.2
python-dotenv==1.0.1
Requests==2.31.0
WTForms==3.1.2
